refactor: domain 구분을 위한 패키지 구조 변경
