+++ src/main/java/roomescape/reservation/dto/response/ErrorResponseDTO.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.reservation.dto.response;
 
 import lombok.AccessLevel;
 import lombok.Getter;
