+++ src/main/java/roomescape/reservation/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.reservation.exception;
 
 import java.time.DateTimeException;
 import org.springframework.http.HttpStatus;
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.dto.response.ErrorResponseDTO;
+import roomescape.reservation.dto.response.ErrorResponseDTO;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
