+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -1,9 +1,9 @@
-package roomescape.repository;
+package roomescape.reservation.repository;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Optional;
-import roomescape.domain.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public interface ReservationRepository {
