+++ src/test/java/roomescape/reservation/repository/ReservationJdbcTemplateRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.reservation.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.domain.Reservation;
-import roomescape.exception.NoSuchReservationException;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.exception.NoSuchReservationException;
 
 @JdbcTest
 public class ReservationJdbcTemplateRepositoryTest {
